+++ src/main/java/com/java6/datn/DTO/ProductDTO.java
@@ -8,14 +8,14 @@ import lombok.Data;
 // trao đổi data giữa client và server
 @Data
 public class ProductDTO {
-    private Long productID;
+    private Integer productID;
     private String name;
     private String description;
     private BigDecimal price;
     private Integer stock;
     private String image;
     private String size;
-    private Long categoryID;
+    private Integer categoryID;
     private String categoryName;
 }
