+++ src/main/java/com/java6/datn/DTO/UserDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserDTO {
-    private Long userID;
+    private Integer userID;
     private String name;
     private String email;
     private String phone;
