+++ src/main/java/com/java6/datn/Entity/PaymentMethod.java
@@ -13,7 +13,7 @@ public class PaymentMethod {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long paymentMethodID;
+    private Integer paymentMethodID;
 
     @Column(nullable = false, length = 100)
     private String name;
