+++ src/main/java/com/java6/datn/Repository/OrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-    List<Order> findByUserUserID(Long userID); // để lấy orders của 1 user
+public interface OrderRepository extends JpaRepository<Order,Integer> {
+    List<Order> findByUserUserID(Integer userID); // để lấy orders của 1 user
 }
