+++ src/main/java/com/java6/datn/Service/UserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface UserService {
     List<UserDTO> getAllUsers();
-    UserDTO getUserById(Long id);
+    UserDTO getUserById(Integer id);
     UserDTO createUser(UserCreateDTO dto); // nhận mật khẩu
-    UserDTO updateUser(Long id, UserDTO dto); // không cho update mật khẩu
-    void deleteUser(Long id);
+    UserDTO updateUser(Integer id, UserDTO dto); // không cho update mật khẩu
+    void deleteUser(Integer id);
 }
